# optimize endianness conversion and tests