--- a/benches/profiling.rs
+++ b/benches/profiling.rs
@@ -7,43 +7,30 @@ use exr::prelude::*;
 use bencher::Bencher;
 use std::fs;
 use std::io::Cursor;
+use exr::image::pixel_vec::PixelVec;
 
 /// Read image from file
-fn read_single_image_all_channels(bench: &mut Bencher) {
-    let path = "tests/images/valid/custom/crowskull/crow_uncompressed.exr";
-
-    bench.iter(||{
-        let image = exr::prelude::read()
-            .no_deep_data().largest_resolution_level()
-            .all_channels()
-            .all_layers().all_attributes()
-            .non_parallel()
-            .from_file(path).unwrap();
-
-        bencher::black_box(image);
-    })
-}
-
-/// Read image from file
-fn read_single_image_from_buffer_all_channels(bench: &mut Bencher) {
+fn read_single_image_from_buffer_rgba_f32_as_f16(bench: &mut Bencher) {
     let mut file = fs::read("tests/images/valid/custom/crowskull/crow_uncompressed.exr").unwrap();
     bencher::black_box(&mut file);
 
     bench.iter(||{
-        let image = exr::prelude::read()
-            .no_deep_data().largest_resolution_level()
-            .all_channels()
-            .all_layers().all_attributes()
-            .non_parallel()
-            .from_buffered(Cursor::new(file.as_slice())).unwrap();
+        for _ in 0 .. 120 {
+            let image = exr::prelude::read()
+                .no_deep_data().largest_resolution_level()
+                .rgba_channels(PixelVec::<(f16,f16,f16,f16)>::constructor, PixelVec::set_pixel)
+                .first_valid_layer().all_attributes()
+                .non_parallel()
+                .from_buffered(Cursor::new(file.as_slice())).unwrap();
 
-        bencher::black_box(image);
+            bencher::black_box(image);
+        }
+
     })
 }
 
 benchmark_group!(profiling,
-    read_single_image_all_channels,
-    read_single_image_from_buffer_all_channels,
+    read_single_image_from_buffer_rgba_f32_as_f16
 );
 
 benchmark_main!(profiling);