--- a/benches/profiling.rs
+++ b/benches/profiling.rs
@@ -9,13 +9,17 @@ use std::fs;
 use std::io::Cursor;
 use exr::image::pixel_vec::PixelVec;
 
-/// Read image from file
+const PROFILING_REPETITIONS: i32 = 1; // make this 100 for profiling longer periods
+
+/// This is a suuuper long benchmark, to allow you to hook up a profiler while running it
+/// but this means we don't want it in our normal benchmarks
 fn read_single_image_from_buffer_rgba_f32_as_f16(bench: &mut Bencher) {
+
     let mut file = fs::read("tests/images/valid/custom/crowskull/crow_uncompressed.exr").unwrap();
     bencher::black_box(&mut file);
 
     bench.iter(||{
-        for _ in 0 .. 120 {
+        for _ in 0 .. PROFILING_REPETITIONS {
             let image = exr::prelude::read()
                 .no_deep_data().largest_resolution_level()
                 .rgba_channels(PixelVec::<(f16,f16,f16,f16)>::constructor, PixelVec::set_pixel)
@@ -25,7 +29,6 @@ fn read_single_image_from_buffer_rgba_f32_as_f16(bench: &mut Bencher) {
 
             bencher::black_box(image);
         }
-
     })
 }