--- a/examples/5b_extract_exr_layers_as_pngs.rs
+++ b/examples/5b_extract_exr_layers_as_pngs.rs
@@ -56,7 +56,7 @@ pub fn main() {
 
         // for each pixel, tone map the value
         for (x, y, pixel) in png_buffer.enumerate_pixels_mut() {
-            let v = data[(y as usize * size.0 + x as usize)];
+            let v = data[y as usize * size.0 + x as usize];
             let v = (v - min) / (max - min);
             let v = tone(v);