from 15 to 7 minutes in ci, nice