@johannesvollmer 2023-07-07
also implements the changes suggested in #193