@johannesvollmer 2023-07-07
msrv would be 1.63.0 in this state