@johannesvollmer 2023-07-07
benchmarks indicate about 10% speed improvement, in addition to the memory usage improvement observed in #193

```
new read_image_rgba_f32_to_f16 ... bench:  36,449,700 ns/iter (+/- 11,014,462)
old read_image_rgba_f32_to_f16 ... bench:  39,865,870 ns/iter (+/- 11,269,662)

new read_image_rgba_f32_to_f32 ... bench:  17,945,880 ns/iter (+/- 5,591,088)
old read_image_rgba_f32_to_f32 ... bench:  20,705,720 ns/iter (+/- 1,138,454)

new read_image_rgba_f32_to_u32 ... bench:  22,919,690 ns/iter (+/- 8,545,283)
old read_image_rgba_f32_to_u32 ... bench:  24,306,690 ns/iter (+/- 1,395,345)

new read_single_image_rle_all_channels               ... bench:  27,453,250 ns/iter (+/- 6,212,354)
old read_single_image_rle_all_channels               ... bench:  29,876,000 ns/iter (+/- 6,462,473)
```